refactor(gui): group device feature modules
