+++ crates/openlogi-gui/src/app.rs
@@ -15,13 +15,13 @@ use tracing::info;
 
 use crate::app_menu::{CloseWindow, Minimize, Zoom};
 use crate::asset::AssetResolver;
-use crate::components::action_ring_panel::ActionRingPanel;
-use crate::components::camera_controls::CameraControlsPanel;
-use crate::components::camera_preview::CameraPreview;
-use crate::components::dpi_panel::DpiPanel;
-use crate::components::light_panel::LightPanel;
-use crate::components::lighting_panel::LightingPanel;
-use crate::components::smartshift_panel::SmartShiftPanel;
+use crate::features::action_ring::ActionRingPanel;
+use crate::features::camera::controls::CameraControlsPanel;
+use crate::features::camera::preview::CameraPreview;
+use crate::features::lighting::device::LightingPanel;
+use crate::features::lighting::standalone::LightPanel;
+use crate::features::pointer::dpi::DpiPanel;
+use crate::features::pointer::smartshift::SmartShiftPanel;
 use crate::keyboard_model::function_row::FunctionRowView;
 use crate::mouse_model::view::MouseModelView;
 use crate::state::{AgentLink, AppState, DeviceRecord};
