+++ crates/openlogi-gui/src/app/detail.rs
@@ -24,14 +24,14 @@ use super::widgets::{
 };
 use super::{AppView, DetailTab};
 use crate::app_menu::file_url;
-use crate::components::action_ring_panel::ActionRingPanel;
-use crate::components::camera_controls::CameraControlsPanel;
-use crate::components::camera_preview::CameraPreview;
-use crate::components::dpi_panel::DpiPanel;
-use crate::components::light_panel::LightPanel;
-use crate::components::light_visual;
-use crate::components::lighting_panel::LightingPanel;
-use crate::components::smartshift_panel::SmartShiftPanel;
+use crate::features::action_ring::ActionRingPanel;
+use crate::features::camera::controls::CameraControlsPanel;
+use crate::features::camera::preview::CameraPreview;
+use crate::features::lighting::device::LightingPanel;
+use crate::features::lighting::standalone::LightPanel;
+use crate::features::lighting::visual as light_visual;
+use crate::features::pointer::dpi::DpiPanel;
+use crate::features::pointer::smartshift::SmartShiftPanel;
 use crate::keyboard_model::function_row::FunctionRowView;
 use crate::mouse_model::view::MouseModelView;
 use crate::state::{AppState, DeviceRecord};
