+++ crates/openlogi-gui/src/components/mod.rs
@@ -1,17 +1,5 @@
-//! Feature-level UI widgets built on gpui-component.
-//!
-//! These are product-specific panels rather than generic primitives. Each
-//! widget owns its local state; cross-widget coordination happens through
-//! [`crate::state::AppState`].
+//! Reusable settings-app UI components.
 
-pub mod action_ring_panel;
-pub mod camera_controls;
-pub mod camera_preview;
 pub mod carousel;
 pub mod device_read;
-pub mod dpi_panel;
-pub mod light_panel;
-pub mod light_visual;
-pub mod lighting_panel;
-pub mod smartshift_panel;
 pub mod status;
