+++ crates/openlogi-gui/src/features/action_ring/editor.rs
@@ -18,7 +18,7 @@ use openlogi_core::binding::{
     RingAction,
 };
 
-use crate::action_icons::action_icon_path;
+use super::action_icons::action_icon_path;
 use crate::state::AppState;
 use crate::theme::{self, Palette, SelectableStyle as _, Typography as _};
 use openlogi_gui::action_ring::icons::ring_icon_path;
