+++ crates/openlogi-gui/src/features/camera/mod.rs
@@ -1,0 +1,4 @@
+//! Camera preview and controls.
+
+pub mod controls;
+pub mod preview;
