+++ crates/openlogi-gui/src/features/lighting/mod.rs
@@ -1,0 +1,5 @@
+//! Keyboard and standalone-light controls.
+
+pub mod device;
+pub mod standalone;
+pub mod visual;
