+++ crates/openlogi-gui/src/features/mod.rs
@@ -1,0 +1,6 @@
+//! Device-feature settings and visuals.
+
+pub mod action_ring;
+pub mod camera;
+pub mod lighting;
+pub mod pointer;
