+++ crates/openlogi-gui/src/features/pointer/mod.rs
@@ -1,0 +1,4 @@
+//! Pointer sensitivity and wheel controls.
+
+pub mod dpi;
+pub mod smartshift;
